feat: Complete dashboard integration with HPLC trace support
Major Updates:
- ✨ Add complete dashboard documentation (README, DEPLOYMENT, HPLC_TRACES)
- ✨ Implement HPLC trace loader for real chromatogram data
- ✨ Add copy-traces.sh script for data migration
- 📝 Create comprehensive action plan (DASHBOARD_ACTION_PLAN.md)
- 📝 Document all architectural decisions (DECISIONS.md)
- 📝 Update root README and NEXT_STEPS with current status

Dashboard Features:
- Real HPLC trace integration from JSON files
- Mock data support for development
- Transformation layer (hplc-loader.ts)
- Ready for Vercel/Netlify deployment

Documentation:
- Complete deployment guide for Vercel, Netlify, GitHub Pages
- Integration summary with architecture diagrams
- Troubleshooting guides and performance tips
- WARP.md for AI agent context

Infrastructure:
- Add proper .gitignore for Next.js
- Include HPLC trace generation scripts (Python)
- Add deployment scripts and foundry config
- Update contract and subgraph schemas

Cleanup:
- Remove outdated crawler and pitch documents
- Archive old description files

Ready for ETH Global Buenos Aires demo! 🚀

+++ contracts/ChallengeManager.sol
@@ -127,12 +127,12 @@ contract ChallengeManager {
 
         if (experimentValid) {
             // Challenger confirms correctness: small positive reward
-            submitterDelta = int256(int64(difficulty));          // +D
-            challengerDelta = int256(int64(difficulty / 10));    // +D/10
+            submitterDelta = int256(difficulty);          // +D
+            challengerDelta = int256(difficulty / 10);    // +D/10
         } else {
             // Challenger flags mismatch: slash submitter, reward challenger
-            submitterDelta = -int256(int64(difficulty));
-            challengerDelta = int256(int64(difficulty));
+            submitterDelta = -int256(difficulty);
+            challengerDelta = int256(difficulty);
         }
 
         reputation.adjustReputation(submitter, submitterDelta);
